add price and dayChange parsing; now works: search by symbols -> find one share -> show ticker, name, price, dayChange in %

+++ app/src/main/java/com/hw/app/database/Share.kt
@@ -9,6 +9,6 @@ data class Share (
     @PrimaryKey
     @ColumnInfo(name = "ticker") val ticker: String,
     @ColumnInfo(name = "name") val name: String,
-    @ColumnInfo(name = "price") var price : Double,
-    @ColumnInfo(name = "day_change") var dayChange : Double
+    @ColumnInfo(name = "price") var price : Float,
+    @ColumnInfo(name = "day_change") var dayChange : Float
     )
